requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.25.2
colorama==0.4.6
contourpy==1.1.0
cryptography==39.0.1
cycler==0.11.0
Fiona==1.9.4.post1
fonttools==4.40.0
geopandas==0.13.2
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.58.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
osmnx==1.4.0
outcome==1.2.0
packaging==23.0
pandas==2.0.2
pefile==2023.2.7
Pillow==9.4.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfiglet==0.8.post1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyproj==3.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-resize-image==1.1.20
pytz==2023.3
pywin32-ctypes==0.2.0
requests==2.28.2
rsa==4.9
Rtree==1.0.1
selenium==4.1.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
termcolor==1.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
urllib3-secure-extra==0.1.0
webdriver-manager==3.8.5
wsproto==1.2.0
rembg==2.0.49
